one more long fix

+++ src/javaObject.cpp
@@ -66,7 +66,7 @@ JavaObject::~JavaObject() {
   jclass nodeDynamicProxyClass = env->FindClass("node/NodeDynamicProxyClass");
   if(env->IsInstanceOf(m_obj, nodeDynamicProxyClass)) {
     jfieldID ptrField = env->GetFieldID(nodeDynamicProxyClass, "ptr", "J");
-    DynamicProxyData* proxyData = (DynamicProxyData*)(int)env->GetLongField(m_obj, ptrField);
+    DynamicProxyData* proxyData = (DynamicProxyData*)(long)env->GetLongField(m_obj, ptrField);
     delete proxyData;
   }
